# common/openapi.yaml
openapi: 3.0.3

# The contract between the frontend and the backend, and the source of truth for
# both: change it first, then the implementations. It lives in `common/` because it
# belongs to neither side — the backend implements it, the frontend generates a typed
# client from it (`npm run gen:openapi`).
#
# Ids in here are opaque: the client gets them from a read or a write and hands them
# back, and never parses or constructs one. Nothing else about how the backend stores
# its state is visible.
#
# Skeleton: domain paths and schemas go in the marked sections.

info:
  title: Vehicle Leasing API
  version: 0.1.0
  description: |
    Skeleton: parties exist, the leasing workflow between them does not yet.

    Every request identifies its acting party through an OAuth2 bearer token, and the
    backend acts on that party's behalf. Reads are scoped to what that party is
    allowed to see, so two users hitting the same endpoint get different answers.

servers:
  # `127.0.0.1`, not `localhost`, and 8081, not 8080 — both for the reasons in
  # `frontend/vite.config.ts`: the backend listens on IPv4, and a published
  # cn-quickstart container on 8080 would answer instead of this app. The generated
  # client overrides this (`frontend/src/api.ts`), but anything driving the spec
  # directly takes it at face value.
  - url: http://127.0.0.1:8081
    description: Local backend

tags:
  - name: Health
  - name: Session
  - name: Parties

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [Health]
      operationId: health
      summary: Liveness
      security: []
      responses:
        "200":
          description: The backend is up
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }

  # --- Session ----------------------------------------------------------------

  /user:
    get:
      tags: [Session]
      operationId: getAuthenticatedUser
      summary: Who the caller is
      description: |
        A 401 means "not logged in" and is an expected answer, not an error — the
        frontend renders the login view rather than a toast. A token naming a party
        that does not exist is also a 401: you can only be somebody who was created.
      responses:
        "200":
          description: The current session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthenticatedUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }

  /logout:
    post:
      tags: [Session]
      operationId: logout
      summary: End the session
      responses:
        "204":
          description: The session is gone
        "500": { $ref: "#/components/responses/Internal" }

  # --- Parties ----------------------------------------------------------------
  #
  # The registry of who exists. Three roles: a car dealer supplies the vehicle, a
  # leasing company finances it, a customer leases it.
  #
  # Two departures from the rest of this spec, both forced by the fact that you
  # cannot be authenticated before anybody exists:
  #
  #   * Unauthenticated (`security: []`). Creating a party is really an administrative
  #     act, so "anyone may create one" is a property of the demo, not a pattern to
  #     carry forward.
  #   * Reads are not scoped to the acting party — the login screen has to list the
  #     roles available to assume.

  /actors:
    get:
      tags: [Parties]
      operationId: listActors
      summary: Every party in the registry, whatever its role
      description: |
        Every role in one list, for the login screen. The only registry read — the
        role types carry no field this one omits, so a per-role list would add
        nothing.
      security: []
      responses:
        "200":
          description: All known parties
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Actor" }
        "500": { $ref: "#/components/responses/Internal" }

  /car-dealers:
    post:
      tags: [Parties]
      operationId: createCarDealer
      summary: Register a car dealer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCarDealerRequest" }
      responses:
        "201":
          description: The dealer is registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CarDealer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }

  /leasing-companies:
    post:
      tags: [Parties]
      operationId: createLeasingCompany
      summary: Register a leasing company
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLeasingCompanyRequest" }
      responses:
        "201":
          description: The leasing company is registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeasingCompany" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }

  /customers:
    post:
      tags: [Parties]
      operationId: createCustomer
      summary: Register a customer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomerRequest" }
      responses:
        "201":
          description: The customer is registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }

  # --- Domain -----------------------------------------------------------------

  # One block per aggregate: a collection path for the reads and the create, plus an
  # `{id}/<verb>` path per state transition. List only the error responses each
  # operation can actually produce — a `404` on a collection read is noise.
  #
  # Two shared pieces to add under `components` with the first domain path, dropped
  # because nothing referenced them yet:
  #
  #   * an `id` path parameter — opaque, taken from a read or a previous write and
  #     never constructed.
  #   * a `404`, for "no such record, or not visible to the acting party". One status
  #     for both, deliberately: which of the two it was is not the caller's business.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth2/OIDC stays in the backend, out of the browser: the frontend holds a
        session and the backend acts on the user's behalf. Until a provider is wired
        up the backend takes the token at face value as the party id, which is why it
        must not leave localhost.

  responses:
    BadRequest:
      description: The request is malformed, or the domain rejected it
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No usable token, or the acting party may not do this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: |
        The write clashes with state that already exists — a party name that is
        taken, or a record that moved on between the read and the write. Re-read
        before retrying.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: Unexpected failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string

    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    Party:
      type: string
      description: Identifies a party. Opaque — derived from the name at registration.
      example: CarDealer::acme-motors

    AuthenticatedUser:
      type: object
      description: |
        Who the caller is. Anything the frontend needs in order to decide what to
        *show* belongs here — but as a display hint only. The backend re-checks every
        request regardless of what the frontend chose to render.
      required: [name, party, role]
      properties:
        name:
          type: string
          description: Display name for the session.
          example: Acme Motors
        party:
          $ref: "#/components/schemas/Party"
        role:
          $ref: "#/components/schemas/Role"

    # --- Parties --------------------------------------------------------------

    Role:
      type: string
      description: |
        Which of the three parties in the workflow this is; exactly one per party,
        fixed at registration. It says what a party *is*, not what it may do to any
        particular record — that depends on the record's own state.
      enum: [CarDealer, LeasingCompany, Customer]

    Actor:
      type: object
      description: A party in the registry, reduced to what the login screen needs.
      required: [party, role, name, createdAt]
      properties:
        party:
          $ref: "#/components/schemas/Party"
        role:
          $ref: "#/components/schemas/Role"
        name:
          type: string
          example: Acme Motors
        createdAt:
          type: string
          format: date-time

    # The three roles carry no fields of their own — a party is a name, a role and an
    # id, with no contact or registration details. Each is `Actor` with `role` pinned
    # to one value, kept separate because the endpoints are.

    CarDealer:
      type: object
      description: Supplies the vehicle and holds it physically until delivery.
      required: [party, role, name, createdAt]
      properties:
        party:
          $ref: "#/components/schemas/Party"
        role:
          $ref: "#/components/schemas/Role"
        name:
          type: string
          example: Acme Motors
        createdAt:
          type: string
          format: date-time

    LeasingCompany:
      type: object
      description: Finances the vehicle, owns it for the term, administers the lease.
      required: [party, role, name, createdAt]
      properties:
        party:
          $ref: "#/components/schemas/Party"
        role:
          $ref: "#/components/schemas/Role"
        name:
          type: string
          example: Northern Leasing AG
        createdAt:
          type: string
          format: date-time

    Customer:
      type: object
      description: Leases the vehicle, drives it, may buy it out at the end.
      required: [party, role, name, createdAt]
      properties:
        party:
          $ref: "#/components/schemas/Party"
        role:
          $ref: "#/components/schemas/Role"
        name:
          type: string
          example: Jamie Fletcher
        createdAt:
          type: string
          format: date-time

    # `name` is the whole of every create request, and not merely a label: the
    # registry slugs it into the party id, which is why the same name twice in a role
    # is a 409 rather than a second row.

    CreateCarDealerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Acme Motors

    CreateLeasingCompanyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Northern Leasing AG

    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Jamie Fletcher

    # --- Domain ---------------------------------------------------------------

    # One schema per aggregate projection and per request body, mirroring
    # `backend/src/api/types.ts`. Each projection carries its `id` as a plain string.
    # Use `Party` for anything naming a party; for anything monetary add a `Decimal`
    # schema — a string with pattern `^-?[0-9]+(\.[0-9]+)?$` — rather than
    # `type: number`, and reference it everywhere, which is what keeps a binary float
    # out of a money field.